todoai-backend/api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "todoai-api"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "todoai-common",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "python-jose[cryptography]>=3.3",
    "httpx>=0.27",
    "alembic>=1.13",
    "python-multipart>=0.0.9",
    "email-validator>=2.1",
    "python-dateutil>=2.9",
]

[tool.uv.sources]
todoai-common = { workspace = true }

[tool.hatch.build.targets.wheel]
packages = ["src/todoai_api"]
